fix:rxns remove bad reactions
For those which are unbalanced in database or contain metabolites which are too general and has no information related, we remove those rxns. (This commit is the first step of updating model for last commit). model = removeRxns(model,{'r_4213','r_4257','r_4259','r_4289','r_4290', ‘r_4182’}); and model = removeUnusedGenes(model) were used in this commit. Rxn ID mapping were based on the massbalance result.