fix-rxn.annot: correct reaction IDs and add new reaction IDs (MetaNetX and KEGG)
New reaction IDs are found during manual curation of reactions in the model. These IDs either replace the current model IDs or are added into the model (if no IDs were originally present).

Inappropriately annotated or deprecated IDs have also been removed.

Details of the changes, such as the rationale for the addition of IDs or changes of IDs are recorded in modifyID.tsv, under the various notes.