fix-met.annot: correct metabolite IDs and add new metabolite IDs (MetaNetX, KEGG and ChEBI)
New metabolite IDs are found during manual curation of metabolites in the model. These IDs either replace the current model IDs or are added into the model (if no IDs were originally present).

Details of the changes, such as the rationale for the addition of IDs or changes of IDs are recorded in modifyID.tsv, under the various notes.